mydiplom/requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.8
django-appconf==1.0.3
django-datetime-widget==0.9.3
django-datetimepicker==3.14
django-debug-toolbar==2.0
django-rest-passwordreset==1.1.0
django-select2==7.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.11
pep517==0.8.2
Pillow==6.1.0
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3.1
redis==3.2.0
requests==2.23.0
six==1.12.0
sqlparse==0.3.1
toml==0.10.0
ujson==2.0.3
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
